refactor: 시뮬레이션 페이지 컴포넌트 분리 및 함수, 변수 명 수정
