+++ packages/client/src/components/simulation/detail/SubjectAllResult.tsx
@@ -16,7 +16,6 @@ function SubjectAllResult({ result }: { result: AggregatedResultResponse }) {
           <th className="py-2 px-2">교수명</th>
           <th className="py-2 px-2">순위</th>
           <th className="py-2 px-2">관심시간</th>
-          {/*<th className="py-2 px-2">관심</th>*/}
           <th className="py-2 px-2">성공/실패</th>
         </tr>
       </thead>
@@ -28,11 +27,6 @@ function SubjectAllResult({ result }: { result: AggregatedResultResponse }) {
             <td className="py-2 px-1">{row.professorName}</td>
             <td className="py-2 px-1">{row.avgIndex}</td>
             <td className="py-2 px-1">{row.avgCompletionTime.toFixed(2) + ' sec'}</td>
-            {/*<td className="py-2 px-1">*/}
-            {/*  <span className={`px-2 py-0.5 rounded-full ${getWishesColor(row.totalCount ?? -1)}`}>*/}
-            {/*    {row.totalCount}*/}
-            {/*  </span>*/}
-            {/*</td>*/}
             <td className="py-2 px-2">
               <span className="px-1 py-0.5 rounded-full text-xs font-bold text-green-500">{row.successCount}</span>/
               <span className="px-1 py-0.5 rounded-full text-xs font-bold text-red-500">{row.failedCount}</span>/
