+++ packages/client/src/components/simulation/detail/SubjectDetailResult.tsx
@@ -14,7 +14,6 @@ function SubjectDetailResult({ result }: { result: ExtendedResultResponse }) {
           <th className="py-2 px-2">교수명</th>
           <th className="py-2 px-2">순위</th>
           <th className="py-2 px-2">관심시간</th>
-          {/*<th className="py-2 px-2">관심</th>*/}
           <th className="py-2 px-2">성공/실패</th>
         </tr>
       </thead>
@@ -26,11 +25,6 @@ function SubjectDetailResult({ result }: { result: ExtendedResultResponse }) {
             <td className="py-2 px-1">{row.subjectInfo?.professorName}</td>
             <td className="py-2 px-1">{row.selected_index}</td>
             <td className="py-2 px-1">{((row.ended_at - started_at) / 1000).toFixed(2) + ' sec'}</td>
-            {/*<td className="py-2 px-1">*/}
-            {/*  <span className={`px-2 py-0.5 rounded-full ${getWishesColor(row.subjectInfo?.totalCount ?? -1)}`}>*/}
-            {/*    {row.subjectInfo?.totalCount}*/}
-            {/*  </span>*/}
-            {/*</td>*/}
             <td className="py-2 px-2">
               <span className={`px-2 py-0.5 rounded-full text-xs ${getStausColor(row.status)}`}>
                 {row.status === APPLY_STATUS.SUCCESS ? '성공' : '실패'}
